[patch] Allow players to vote when they think a node is having issues.
Votes will be valid for 10 minutes,
if 50 of playing players vote for an issue 1 of 2 things will happen.

1. if an external node is the current node then within 15s of it reaching 50% of votes will trigger a reconnection

2. if its the managed node that received 50% of the votes then a node restart will be triggered.

in both of the above players will be migrated to other nodes if you have multiple, or they will wait for the node to become available again.

the api is available via `Player.vote_node_down` - the action can be reverted by calling `Player.unvote_node_down, every subsequent vote resets the 10min counter for the player - but duplicate votes are ignored (Each server is entitled to 1 vote, if they are actively playing)