[minor] Make 0.9.0.0
This release has a lot of changes to the config backend, namely the ability to enable/disable caching via the `PYLAV__CACHING_ENABLED` variable, if this is enable PyLav will cache the postgres queries to reduce the number of call - probably useful if you have a remote server and without it you are noticing performance issues.

Query and Playlists are now also stateless to follow the changes introduced bt 0.8.x

The radio browser cache will be primed on library initialization to improve performance for when it is used for the first time (as it can take anywhere between 4-10 seconds for the full query to return a response

Added for players to report issues with their current node, if 50% of more of players report an issue a node reconnect (if external) or restart (local managed) to the node migrating the players as necessary.

Add a `Player.ping` attribute which represents the gateway ping in milliseconds

All config modals are now singletons and have their cache stored within them.

PyLav objects should now be more memory efficient as well as more performant due to the usage of `__slots__`